backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-login>=0.6.3
flask-session>=0.5.0
pymongo>=4.6.0
flask-pymongo>=2.3.0
yfinance>=0.2.28
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
statsmodels>=0.14.1
tensorflow>=2.16.0
bcrypt>=4.1.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
cryptography>=41.0.0
gunicorn>=21.2.0
